requirements.txt
pydantic==2.9.2
fastapi==0.115.3
SQLAlchemy==2.0.36
alembic==1.13.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
asyncpg==0.30.0
bcrypt==4.2.0
pyjwt==2.9.0